+++ packages/is-shallow-equal/test/index.ts
@@ -146,7 +146,7 @@ describe( 'isShallowEqual', () => {
 	} );
 
 	it( 'returns true on arrays that are a copy of each other', () => {
-		const a = [];
+		const a: unknown[] = [];
 		const b = a;
 
 		expect( isShallowEqual( a, b ) ).toBe( true );